requirements.txt
fastapi
uvicorn
pyodbc
sentence-transformers
faiss-cpu
numpy
openai